fix: resolve issue #563 (#708)
